feat(abstract-lightning): move lightning wallet update
to self custodial

Ticket: BTC-0
